Update 0.8
**For now I have given up on changing the size of the game. It would
really only require me to change the size and arrangement of UI
elements but I want it to be built with the iPhone screen in mind.
Thus, in order to play the game via web browser I recommend zooming out
to 50%.**

-Canon is now operational. After destroying a hazard with the CLAW,
players can press the spacebar to fire a cannonball and then must wait
until the CLAW recharges and can destroy another hazard.
-Created a battle mini-game that is extremely unlikely for players to
win, as intended. Can be fairly easily modified to change the numbers
and names of moves.
-Battle mode involves selecting a move and taking/dealing damage
according to the randomly chosen enemy move. Enemies start off with
wider array of moves.
-Battle screen needs to be made prettier, attach lore and make the
numbers “pop”.
-BattleManagers needs additional variables to determine the enemy’s
name, sprite, eat…
-Added UI to the battle screen to indicate damage taken and moves used.
-Battle mode updates health for player.
-Added a restart option upon failure.

Known Bugs:
-If the player is killed by hitting two objects at once, the screen
continues to shake.
-Players can continue taking moves after dying in the Battle mode.