Update 0.9
-Reworked a ton of the code for the enemy prefabs. Enemies now share a
couple of common scripts for their movement, firing mechanisms, ect
instead of each having an individual script with the same code.
-Played around with lots of the numbers for prefabs, very distinct play
styles can be created just by tweaking the numbers.
-Added an enemy called Chaser that is the same as a Wagon, but instead
of rolling downwards it trots alongside the player. Also fires 3 shots
at once at a slower pace.
-Added code to the Hazards script so that they now damage enemies.
-Added enemy health system.
-Created sorting layers and added them to all necessary prefabs.

-Created the village of Laddingtown based off of Verenfield’s UI
system. Wrote the lines and arranged the dialogue chain to be fairly
straight-forward. Used the Fallout 4 trick of having multiple
conversation tracks that ultimately lead back to the main path. I may
rework Verenfield using this strategy so that I can be sure users
actually see the main story and figure out why they’re actually doing
all the shit they’re doing.

-Fixed an initial bug where enemies would fire their shot and instantly
take damage. Created a boolean that is set to true after a .5 delay.
Lead to a minor bug that if bullets are fired very close to the enemy
they will simply roll over them (because they have colliders have
already entered each other). I currently like this but would like to
see the player be able to utilize timing and strategy to quickly turn
the enemies bullets against them.
-Fixed a bug where enemies everywhere were firing when the player
entered the range of another enemy. This was due to me creating a
static variable so I reworked the code to resolve that.

Known Bugs:
-CameraShake is still causing issues. Minor, but when killed by two
hits simultaneously the camera continues to shake instead of pausing,
as intended.
-When the Sweepers lose their sweepers there is a null reference
exception. But I likely am going to rework the Sweepers so it may be
irrelevant.